+++ src/file/mod.rs
@@ -12,3 +12,5 @@ mod track_event;
 
 pub use division::Division;
 pub use header::{Format, Header};
+pub(crate) use track::ensure_end_of_track;
+pub use track::Track;
